refactor: remove AgeController. Move the Age package to Timetable.
