+++ src/main/kotlin/odyseja/odysejapka/timetable/PerformanceEntity.kt
@@ -1,6 +1,5 @@
 package odyseja.odysejapka.timetable
 
-import odyseja.odysejapka.age.AgeEntity
 import odyseja.odysejapka.city.CityEntity
 import odyseja.odysejapka.problem.ProblemEntity
 import odyseja.odysejapka.stage.StageEntity
