+++ src/test/kotlin/odyseja/odysejapka/timetable/ImportCsvTest.kt
@@ -26,7 +26,6 @@ class ImportCsvTest: OdysejaDsl() {
 
     @Test
     fun `should import csv file`() {
-        ensureAgeAndProblemExist()
         val content = this.javaClass.getResourceAsStream("/import-csv-test-cases/fo2025.csv")
             ?: throw IllegalArgumentException("Brakuje pliku fo2025.csv.")
         val csvFile = MockMultipartFile(
@@ -63,8 +62,6 @@ class ImportCsvTest: OdysejaDsl() {
 
     @Test
     fun `should reject imports with improper data`() {
-        ensureAgeAndProblemExist()
-
         val testCases = listOf(
             mockCsv(performanceDay = "wtorek") to "Dozwolone dni występu to sobota lub niedziela.",
             mockCsv(spontanDay = "2") to "Dozwolone dni spontana to sobota lub niedziela.",
